Update issue 1217
Was able to get values for the harvested energy output to the shapefile.

However, there is a question over whether these values are correct and whether I
have set up the integration properly. Ask Rich about this when he returns next
week. I should be able to continue on regardless.

Also had to rename the harvested wind energy function because it was the same
name as another variable I use and that was causing conflicts


Former-commit-id: 60e833dcada8db911ba4ac1b0e5240703709b263