Update issue 1217
Attempt to vectorize the points onto a raster. I am using vectorize_points from
raster_utils but have run into an error. The error I get states that the array I
am trying to write out to the band is not the same size and is to big. After
debugging I found that my raster has a band shape of 73 x 110, and the array i'm
trying to write is 74 x 110. Thus I think something might be slightly off in
vectorize_points function.

I am creating my raster from the points shapefile that I am trying to
interpolate. I think at this point I will wait for Rich and continue on like it
was working


Former-commit-id: b09036b395838cb260a0fab07dadc6ba471a7a58