Update issue 1172
Did a huge overhaul to how the bathymetry dataset is handled. I had a function
that handled this, but realized that it was just lines of code making calls that
could be easily done in the maine execute block. Basically it was not reducing
code or duplication issues. Therefore I felt like the function was non worthy
and did all the handling for clipping and reprojecting the bathymetry in the
execute block.


Former-commit-id: d95981be9b8688bd0f71d2a02d8987d4c49dcedf