New issue Summary: Write HRA Model
Create skeleton for HRA core functionality. This will include the equations for calculating exposure and consequences for particular habitat-stressor pairings.

Should be noted that we NEED to figure out what the data structure for this is going to look like. In terms of data needs, it comes out like a 5 layer tree, or a 5D array. I understand that all of this must be pulled from a CSV, but that's not axactly clear to retrieve from, unless ywe're using a dictReader and creating a 5D dictionary.


Former-commit-id: c2132e2042d19bd323e94523c875cc7a291981b7