Update issue 1230
Docstring for calculate_exposure_value. Have not yet decided what will be passed as params into the function.

At the very least, we will need an iterable for the criteria. For now, just a "for each" loop and assume that whatever we get as a param will be iterable within a for each loop. Can decide later the exact details.


Former-commit-id: 3064bb04f8c0cfdb843abe1821d9de9366a4cd23