Update issue 1230
Created function for calculate_exposure_value.

For now, I am imagine that there is an iterable structure containing the criteria that can be accessed by iterable[criteria], within which is contained a tuple of structure (value, data quality, weight).

This will be edited later, but the equation itself is sound.


Former-commit-id: 13bca7d3403799d07bb14f288d1c038d3493849b