Update issue 1236
Created a rough user interface that compiles correctly. With James's insite I'm
thinking of having the type of turbine and relative information be allocated in
a CSV file. We can thus provide the two default types and CSV files and if
people want to customize they can simply copy and edit one of the current ones.
This simplifies things greatly


Former-commit-id: b9095cab9602b10c9bed03d6830e1087b34f79bb