Update issue 1229
Finally found the problem with erroneous land_elevation data: it's a gdal bug. The solution to this problem is actually more elegant than the initial one I came up with. Instead of overwriting an existing raster and change its nodata, I now combine two shapefiles to create a new one, which is the combination of the two. This way, the effect of the function is more explicit, I don't have the side-effect of changing a file on-the-fly, and the code around it refactors more elegantly.


Former-commit-id: e283a1d38c1f56a240a440a7e29e3a6d84c71a56