Update issue 1230
Created the skeleton for the ecosystem raster. Instead of basing this off of the allready open datasets that I have within the ratings structure, I'm going to retrieve all datasets from the folder into which they were rasterized.

However, also teh option of returning a value from the create_habitat_raster_function, which would be a list of all the open habitat datasets. This could then be passed directly into the function that will combine the habitat rasters, but at the expense of some memory, I believe.


Former-commit-id: feacca5b0287c528313e7f274dbbb31b0a369f24