Update issue 1276
It looks like my old school way of interpolating points onto a raster is causing
it to bust. I therefore switched it out for Rich's vectorize_points in
raster_utils.

I ran it on a smaller dataset and the outputs still loook good. Going to push to
James, so we can give it a whirl on Windows


Former-commit-id: adbdb4201da223eac71d2b6e3b900c67ceea2ea0