Update issue 1261
There's a discrepancy between the size of the shore and the population density rasters. The shore is rasterized directly from the AOI projected in meters. The population raster is in lat/long, and is clipped to the AOI, reprojected in lat/long. When I project the clipped raster from lat/long to the AOI's original projection (UTM 10/WGS 84), its size doesn't agree with the shore raster. In short, these two pathways:
AOI shapefile (UTM-10) --> rasterize
data raster (lat/long) --> clip to AOI (lat/long) --> project to UTM-10

yield rasters of different sizes.


Former-commit-id: 4f1a170764fc5ed2c8f29283b4bfe8ce38801953