Update issue 1261
The size of raster produced by the function vectorize_raster depend on the order of the rasters in the list. In one call, the composite raster size is the smallest size of the rasters in the list. In another call, the raster size is the largest of the rasters in the list, minus 1.


Former-commit-id: 403f9a3f448a198a84ba5d1ffc9c11c39700c262