Update issue 1230
Started to set up structure for making the recovery potential raster, but realizing that it want to know what the predominant habitat is for every cell.

This is fine for something where there are differences in amount of cover from each habitat, however, what about pixels where there is the same (whole or none) coverage from multiple habitats? How does one then select the predominant habitat for that cell?


Former-commit-id: 57e83d56ff4c71cbb6ed53bb24706c28b3aad828