Update issue 1258
Adding some rough draft code for vectorizing points for the ocean to land and
land to grid fields in the point shapefile. Thinking that if I make these into
rasters then I can just do a nice vectorize op and use vectorize_rasters to
compute the NPV.

Is this the most efficient way, or would it be more efficient to iterate over
each shape point and do the calculations, then vectorize points?


Former-commit-id: 502b91be6b8446b0aa9e2ec80ccf8bec0e6536dc