Update issue 1245
Take the array of the stressor raster, and create a numpy array where each value is the distance from "land," and "land" is represented by 0. Then, turn everything within the buffer zone to data values, and everything outside that to nodata, and write it back to the stressor raster to be combined with the habitat rasters.


Former-commit-id: bb6bc509f1a0278d0d67a0c92a0852d209122892