Update issue 1289
Before using the guassian_filter we now read in the raster once, save it to disk
and then use numpy.memmap (memory mapping) to load back in the data to
manipulate and use.

Also added comments

I ran it on my linux machine on the user's large data and I got through it all
with no errors and no memory spikes on my machine. The outputs look good.

The changes do not affect the unit / regression tests.

This should now be pushed to James who has a Windows 32 bit machine for proper
testing. Cross our fingers.


Former-commit-id: 56b4e84987e1848622fc84a60679838188140600