Update issue 1248
Wrote single unittest for standard run of the HRA model. Should just run a basic version with some of the criteria existant, though some have 0 weights, and the biotic are not added. These can be added later as we figure out how to differentiate between biotic and abiotic areas. However, the ratings only has one pairing for H-S. Need to copy these two to another set of folders so that we can do a minimal test without having to write an overwhelming ratings structure.


Former-commit-id: 9f09f27ea58796103e4cce65e68f4f0657e282a1