Update issue 1230
Changed the E/C equation into one function calculate_score_value, instead of forcing them to be two identical functions.

Additionally, now if a 0 is passed in either the data quality or as the rating of the criteria itself, the addition of that criteria to the E or C sum will be 0 overall. This should hopefully help us avoid the divide by 0 error.


Former-commit-id: 3823669d51545bac9e7d9111e3cec53fbe289c60