Update issue 1230
So, I'm going to push this, but it shoudl be kept in mind that the way I am going about this is wrong. What needs to happen is that there will be a different recovery value of sorts fo each H-S overlap, then all of those will be combined. Still waiting to find out if DQ/W are going to be included somehow in these calculations.


Former-commit-id: 83b13e721a0c723d54a05e02bf84281560b9bceb