Update issue 1324
Adding an entry to the invest setup.py so that the new nutrient package will be brought along into the built virtualenv.  Also, adjusting the import statements in nutrient_biophysical_test and changing the nutrient_biophysical execute function to print the input args dictionary for the time being so I know that we got there.


Former-commit-id: 3006f69077c93d01131e18026628620f82451645