Update issue 1317
Did some major overhaul on how the distance mask is created in the core. Using a
similar setup to the gaussian_filter_dataset function in raster_utils

However, there seems to be a bug with setting back the nodata values using the
following method:

dest_array[mask_array] = out_nodata

it doesn't seem to work.
mask_array.any() prints True
the shapes are the same
out_nodata is a unique value


Former-commit-id: cb859bf5837837c1457dfe2a6f165cf4645fbad8