Update issue 1301
Fixing a bug with the disk space logging on linux, where the disk space function was crashing if the folder did not exist.  This exception was then sorta-caught by the exception handler that called it and no error was reported.  Not the greatest problem in the world.

I still need to test this on windows, whether providing a folder that does not exist to the function will likewise cause it to crash.


Former-commit-id: 2b3992e15f40a2d011a95c62b38d2b844b473798