Update issue 1230
Setting up basic skeleton for the make_raster_files function. We will pass 3 dictionaries to the E/C calculation function- the main h_s one for the E or C ratings, and then the E or C for both applicable habitat or stressor.

When we go to do risk calculation itself, we will pass the unmodified h-s overlap dataset to the euclidean risk or multiplicative risk function, and have them spit back out a dataset which properly calculates the risk of the h-s overlap.


Former-commit-id: e5c11b4a383c9472bfd35e4940ea293cd4586ee0