Update issue 1230
Wrote method to return score value for either E or C based on the three dictionaries that should contain criteria scoring. Fuction gets passed subsections of the three dictionaries, all of which should be identical, and then adds all the scores to top/bottom, and dives the two at the very end.


Former-commit-id: 11e4203ff3d1e51acc2aefe406cbc02aa7655fc8