Update issue 1230
The call to the "make risk raster" function will actually take in an array of the raster band, and return a modified array which will be burned back to the band. In addition, it will take in E/C and will use those to apply either using a multiplicative function or a euclidean distance function.


Former-commit-id: ef02f26fdc655d60df3b020aa3ff0c7594494053