Update issue 1307
So it seems like by far the easiest way to do this internally will be to calculate everything and then save it to the correct column of a given shapefile instead of doing things in tables with records that correlate to the user's defined watersheds and subwatersheds.  To help facilitate this, I'm now creating an ESRI shapefile copy and passing that in the biophysical_args dictionary.  Still need to create new fields, though.


Former-commit-id: ebf8d7dcedbac5572690f0e54094205ec94cabf9