Update issue 1307
In consultation with Rich, removing the new rasters created by the model.  These rasters are redundant because pixel-scale values are not the correct way to visualize watershed values.  Instead, I'll still create the table outputs created by the InVEST 2.x model but I'll also output the watersheds shapefile with the values in the attribute table.


Former-commit-id: 3d89c0bfa76d18043938abbf63afbab31f9751cf