Update issue 1308
The numerical intervals spanned by E_o and E_l are one order of magnitude apart. This is due to the effect of the fetch in equation 6 (E_o) as opposed to its absence in equation 0 (E_l). This makes sense, since the power of oceanic waves can be much stronger than local wind-generated waves. For this reason, it makes sense to do the ranking on E_w itself, even if it contains a large numerical variations. ALso, the ranking will be done in average-sized bins, as discussed with Greg on October 16-17th.


Former-commit-id: 79c715e6ac6071a8f131e7ffad2165ef88094eb8