Update issue 1230
Continuing to finish recovery potential function. Realizing that in order to re-rasterize the habitat raster with the updated recovery potential value, we will need to have the open raster datasets for the habitats.

Adding into docstring that we will be passing a 'DS' value into the args['habitats'] raster in much the same way that h-s contains an open raster dataset.


Former-commit-id: 8839bc9a6d46cb029cae8b2cb6051790f08fb848