Update issue 1335
So Rich suggested using GDAL to do calculation of raster statistics (and then use the calculated mean therein) rather than calculating the mean through numpy.  I was skeptical of the runtime improvements, so I wrote a script to check this.  Currently lives as a commented-out test, though I should really move this to a standalone script for fun.


Former-commit-id: fac8b54dd7e9acee14364fafba6df2f5b2c8124e