Update issue 1230
Actually finished the recovery_potential function. It pulls the band as an array from the original habitats raster, and multiplies it by the calculated recovery potential value to keep the 0 nodata values.

Should now output a raster with the same land dimensions as the original habiat raster, but burned with the recovery potential score for that particular habitat.


Former-commit-id: ebca1c6deb17ffbe5721617a44e1aa22c9cf47bf