Update issue 1230
Changing the risk_raster function to create a new risk raster file per H-S overlap, instead of re-burning the original. We are creating a separate dictionary to contain only the H-S risk rasters that will be returned from the function, and can be passed on to the habitat raster creation.


Former-commit-id: 5707fcce2c2822d5f532114280cac60e2e36c877