Update issue 1245
Changed the calls which previously used the re-rasterized versions of the buffered stressor shapefiles. Instead, we now want to create new versions of the files, and add them to the stressor raster dictionary which will then be passed to the function creating the h-s overlap.

If the desired buffering for the given stressor is 0, the "buffered" version of the shapefile will be identical to the unbuffered version. We are still placing the buffered files in the same Intermediate/Stressors folder as previously. However, since how we will have a copy of the file which says "buffered" we will just call the buffered version of the file, knowing that it has already taken 0 buffering into account.


Former-commit-id: 0dfd7291b56605453fe6e3afc2e70672d20b0893