Update issue 1245
This will create a new file containing the buffered stressor files. We still need to take in the user's prefered form of decay and make sure to get a proper buffer decay, but we can just create individual functions to do that (much like for risk) and pass and receieve back numpy arrays.


Former-commit-id: 2587487682657256ebf5e2b402f535e2fec39b37