Update issue 1335
So realized that my previous comparison of the wall-clock running times of calculating statistics on a GDAL raster vs. a Numpy array was inaccurate because I was only calculating the mean in my numpy tests, while GDAL actually calculates 4 things: the min, max, stddev AND the mean.  So I added calls to get the min, max and stddev in each of the numpy timing tests.


Former-commit-id: fa6db067513b9035733d1936d4f44fe53c34408f