Update issue 1236
Slight modifications to user interface.

Thinking that I will only need the biophysical wind data points as an input and
not the harvested raster.

Also trying to think of the best way to handle the distance issue. The user
should have the ability to either enter Grid data points as a CSV file or a land
polygon. Can one do OR in the UI?


Former-commit-id: 5a0dacb18512fcceec3f9d7105838378d7c14f6a