Update issue 1319
Commenting out the added handler in the get_ui_logger function.  This handler addition was causing all logging (even model logging )to be printed twice to stdout.

Also, commenting out the one line in base_widgets that was using a call to LOGGER.  This way no UI logging should exist, so there shouldn't be anything extra printed to stdout.


Former-commit-id: 6f65c9d9025886321f0342841c44196dd0ac997f