Update issue 1335
Adding a new function get_retention().  This function's purpose is to get the amount of nutrient retained by the landscape per the table in the user's guide (you know the one I mean).  This function's guts will eventually be replaced by a call to the comutional framework that Rich will provide an API for in the near future.


Former-commit-id: 06f1c1efd5a949abf95e64905ba81de5fb3fc6c3