Update issue 1382
Remove create_param_log function, and function call, since the IUI now does that for us.

Should be aware that the current IUI does not display any args coming in through the CSV. If the user does want those, we will need to either output them to the LOGGER, or somehow re-create the self-called param log.


Former-commit-id: d5989eef9cb128ef25be7a61477287918b5b3583