Update issue 1250
Added a function to clip and project a datasource. This turns out to be super
handy and useful since there is a pattern for this process of projecting the AOI
to shape_a, clipping shape_a, then projecting shape_a.

Adding this function allowed me to shorten up a lot of code.


Former-commit-id: 73e23405750d824da90278c01fd8c07b069249fe