Update issue 1286
Refactored coastal_vulnerability_test.setUp to split the variable initialization code (input from UI) from the rest. Fixed a bug in setUp that overrides the aoi for the area of Vancouver Island as a raster and causes the code to break. Fixed test_adjust_layer_ranks so that it passes.


Former-commit-id: 66010badd39b2c88d92d1a43b68cfe906243bd35