Update issue 1251
Added a ton of code to handle either grid points or land polygon to compute
distances. The code is currently super scattered but it runs as expected.

Added a new function to return a dictionary from a shapefile where the geometry
is it's key and the sub dictionary it points to is the fields and values.

Now using kdtree to get relationship of ocean points to land points.

The code needs to be super cleaned up.


Former-commit-id: 5da3797c6bc4c18ba4aa025efc54006c3667c4a5