Update issue 1251
Added the vectorization of the points for the NPV field.

After viewing the otuput an interesting problem occurred. The Raster is
obviously made from the vector extents of the points, but it has nothing to be
clipped against for a meaningful output. I'm starting to wonder if perhaps we
need to rethink the inputs.

For example, could we make the AOI just cover the ocean points / area you are
interested in, then also make the land polygon required so that we can mask out
any land values that the output raster might interpolate over.

Also wondering if it is necessary now to have the AOI cover a piece of land. I
think we are using some pretty quick efficient functions for computing
distances and such, where my worry before was that we would be calculating
distances from a point to every polygon in the world (is this a bad thing?)


Former-commit-id: c8dc796d5c39aef2fbf9bc958e2b18589cc0a2e5