Update issue 1394
Adding some more logging and fixing a bug in my checking of isinstance.  Exceptions raised in the model are now passed on to the operation dialog and not just killing the thread.

So the underlying issue here is actually that the recretion_client script is raising an exception that it cannot find invest-natcap.invest-3/config.json.


Former-commit-id: 3b1b13ae01f90dfa2b3fd3d1c12a2d645a921634