Update issue 1230
Keeping track of the number of pixels which overlap out of the total number. Will divide to get a percentage, then enter in the spatial overlap scores based on the percentage. Will have to explicitly define which percentages correspond to which scores.

Need to find out what the deal is for spatial overlap's weights and DQ scores, since we aren't manually entering them.


Former-commit-id: 926c51f321aee8a870fa50193abdb890d03bfd84