Update issue 1230
Now calculcates the percentage of overlap based on the overlap pixels/ total pixels. Runs a series of if statements to determine the proper rating based on that percentage. (Values pulled from the user's guide.)

Should be noted that I made up values for the DQ/W, since I don't know what the "default" for comp-generated criteria should be.


Former-commit-id: b3144cde7167fa32e3912a59db880122d3248a5d