Update issue 1283
Modofied cast_ray so that it can compute the average depth along a ray. This information is going to be necessary to generate the following outputs:
        -a: exposed and sheltered shoreline segment map,
        -3.4: Surge estimate,
        -3.5: Wind-generated waves.

The modification works and the existing tests still pass. Need to document the change in the docstrings, and develop a test to compute the average bathymetry.


Former-commit-id: 81603524d37a5959489a460abdbb3b02c1d66686