Update issue 1373
Adding some more logging information to the invest-logger that will hopefully help us to get a better sense of what sorts of systems our users are running InVEST on.  New information logged includes the OS name (e.g. 'Linux' or 'Windows'), the OS release number (such as '7' for windows 7 or the kernel revision if on Linux), the foll platform string and the filesystem encoding scheme.


Former-commit-id: 6159646ac14e8fb5221a990819a730ad1aa0d608