Update issue 1230
Made the recovery potential equiation dynamic. Put in a check to remove the criteria from the equation if either the data quality or the rating itself were 0. Should never have an input that is null, since we should be checking at dictionary inception that everything is a '0' value that needs to be explicitly called.


Former-commit-id: 356c36f61a70eebcdc223f22f475eefdeefa22a6