Update issue 1404
There was a place in new_raster_from_base where a URI was being cast to a python bytestring (str).  This is fine if the user is running it in the US, but if you're trying to cast a different encoding to a bytestring, it will raise some Unicode-related errors.

To fix this, I just changed the cast from a bytestring (str) to a unicode string (via the unicode constructor).


Former-commit-id: d6223b60f27e69e6b5236ed073615fd646fa8624