Update issue 1347
Adding a function to setup.py to write the version number to a file.  The file that is written conforms to python syntax, including comments and a variable that is set.

This python file is then imported in invest_natcap.__init__.py and used if and only if the version is hand-set to 'dev'.


Former-commit-id: c9ef8a5d4ee53025a847cb76a691f15354536ac9