Update issue 1347
Reworking the invest_version() function.  Now, if the invest version has not been generated for the current state of InVEST, one will be created on the fly and the correct version number will be returned.

This approach is a little risky because it is dependent upon setup.py always bringing along all python files with any sort of installation.  That being said, this functionality is such a core part of setup.py that I think we can reasonably say that we will always distribute InVEST with the generated invest_version file in it.


Former-commit-id: 4d04cc9e3a587fa990b14f2c656cf04ebbf460a1