Update issue 1347
Modifying build_utils.invest_version() to have an optional parameter to write the version/build information to disk.  If True, this information is written to disk as it did before.  If not, then the correct information is returned but the information is not written to disk.  This argument defaults to False, so if someone just imports from the source tree, hg will be called to get the correct version/tag information.


Former-commit-id: dcf018abc8a748cfd573a39bc96804bc6a813047