Update issue 1418
Nutrient_valuation now makes a new copy of the input watersheds
shapefile in the workspace.  I know this is a little redundant, but it
at least ensures that _if_ the user uses a separate workspace for the
valuation step from the biphysical step, the watersheds shapefile is
saved to the valuation workspace.


Former-commit-id: 907faea5ff5c21e96d417f208b0cfec9d70cbb00