Update issue 1423
Adding functionality for the user to remove the lastrun ('cached run')
file for the currently running model.  To do this, I added a function in
the lastrunhandler class to delete the underlying lastrun file if it
exists.  The lastrun dictionary (the internal dictionary tracking the
lastrun values loaded from the latrun file) is then cleared.

This functionality can ben invoked through a new option in the file
menu.  Note that this strictly clears the cached values and does not
reset UI values as well.  If the user wishes to reset values, they can
click the 'reset' button.


Former-commit-id: cd88dc11e047284b7a3892eb60b14e2ec5e9453c