Update issue 1404
Trying a different encoding scheme in the new_raster function in
raster_utils.  Now, all strings input to new_raster() are encoded to
utf-8.  This means that if a string is already encoded to some unicode
codec, a utf-8 unicode string will be returned.  If an ascii string is
passed, it will be encoded as ascii (since utf-8 is maximally compatible
with ascii).  Since GDAL is happy with either ASCII paths or Utf-8
paths, this should be a good solution.


Former-commit-id: b051917cf2c91b973fe05f56c9015acd3cebc3d9