Update issue 761
Adding logging to the pollination test file.  This is used at the moment
to verify which regression rasters are being compared.  Part of the
problem with the regression test in issue 761 is that I was accidentally
comparing the regression raster against the regression raster ... not
the raster that was being created by the model.

This commit also fixes the issue of actually comparing the two rasters
correctly.


Former-commit-id: bcc0e6d84b6a65176001d2c95983c46c7033927a