Update issue 1494
Re-factoring the Biophysical URI to handle the changes in how the biophysical
parameters are coming in as inputs. Now loading / reading from a CSV and JSON
file for the inputs.

Still need to see these edits through to the core level because the core is
note expecting a dictionary of values.

Also this must be done on the valuation side as well


Former-commit-id: 46c65a35c5df569273ac0c96c4e9b6c2d4781f3e