Update issue 1494
At the moment there is not validation for the CSV file or JSON file. I thus
added a list of what fields / parameters are required and check that they have
been added to built dictionary by comparing lengths.

I then make my own exception class and raise an exception if the lengths do not
match.


Former-commit-id: 5b46f485911dc8ce043930d90793946ae04da5b1