Update issue 1513
Modifying reclassify_by_dictionary so that nodata values are preserved
if a default value is specified.  This means that a default value is
only returned for a pixel if the pixel is not nodata and the pixel
doesn't have a reclass rule.

This behaviour can be overridden by the user by defining a rule for
nodata values.


Former-commit-id: 5e8fd7cd254d6ae5cdb9fa9bae3029e637a24554