Update issue 1490
Currently I was dumping all the information from the wind data text file into a
dictionary and just adding everything as a field to a shapefile. Looking over
how I was reading in that text file, I realized it made no sense and have thus
refactored how the function 'read_wind_data' works. It now reads in the file
like a CSV using DictReader to parse through the rows. The function also takes a
list specifying which fields to add to the dictionary. This way, we only add
what we really care about and what we want to add to the shapefile.

I next need to get the hub_height and create the proper field name from it which
I currently do in the core. It now has use here.


Former-commit-id: 6a163a61395954f697733eb8cac125c0e8fea2ef