Update issue 1290
Now calculating the farm_value_sum raster using vectorize_rasters.  This
raster is currently missing the necessary URI, so I'll need to figure
out just how to handle this.

I think that the best way to handle this would be to take the ag_map
(which is based off the LULC) and reclass all non-nodata values to 0.0
and then have the URI provided as an args input.


Former-commit-id: 9d9dd14881dfbd2d11e7690e0d4db6c54cae16ff