Update issue 1290
Adding a test that should be skipped by default.  It tries to run
pollination  (all of pollination) on a 1m-resampled version of
lulc_samp_cur.  I tried running it on my office computer, but it took up
way, way too much disk space (as in, all of it) before it crashed from
running out of disk space.  I'll have to try running it on my office
windows computer, which has much more disk space.


Former-commit-id: ef3140b25db83885981b2f5924ad1f6e0aff2822