Update issue 1290
Adding a function skeleton and a docstring for a function to calculate
the service value raster.  This raster has a rather large number of
inputs, but it at least performs a single discreet operation (and all of
the intermediate operations leading up to that one).

I anticipate finishing this function as part of the pending merge of
pollination biophysical and valuation into a single tool.


Former-commit-id: 68db74d3cbe1a344a4dc80f8db873b480da7a130