Update issue 1251
When creating the raster outputs at the end of the run, I wasn't properly
closing the datasets and am afraid they were just left to Python's garbage
collection.

I thus flush and close them now to be extra safe. Because after these are
created I use one of them to get some info from and since I didn't have a handle
on it I was worried that re-opening a file that was open might cause a fuss


Former-commit-id: 27c057d327061c699c2706b11027cba76ceb1a08