Update issue 1230
Ridiculous large commit. Rejiggered the E calc so that it pre-rasterizes each criteria, then adds them to a list to be vectorized a second time with all of the other E-specific rasters. Still need to do the same thing for C calc, except over two dictionaries, rather than just stressors.


Former-commit-id: 126137545f5e2434d63d975ff1d006977388374d