Update issue 1230
NEW STRATEGY. Pre-calculate the 3 criteria raster versions ahead of time. That would save SO MUCH visual complexity when trying to read through the code. Then everything will just be a matter of iterrating through a few more dictionaries. Additionally, want to set up the denominators dictionary at the same time.

ALSO: Should enable us to use the same code w/o an if/else for the spaital criteria, since "everything" in the list will be iterrated through. If that item is just the ratings raster, that will be unaffected.

Former-commit-id: 6357c6b0971417a79cd10a7bcd70fb77fdb76c9a