Update issue 1356
Test code is in place, and executes, but I'm having a problem with the precision: when I save the fetch fata to a csv file, the floats get truncated. When I load the truncated data and compare against the exact same data (which has not been saved), therer can be disagreement up to the 5th decimal, which is unacceptable. The precision loss happened when I save and reload. Looking into this.


Former-commit-id: a1b1866f84258d42cc70405ab63bc1361a4bef9c