Update issue 1356
The test on the regression data fails because the rays don't contain the same depth information, as if the first element is duplicated in the fast version, but only sometimes. What's perplexing is that the average depth should be off by a about 20%-40%, but it's only off by 10^-6 or 10^-7. I think the sum is taken so that it skips the first element and shoots off one element too far, which is an undefined value of about 10^-7. This makes sense, because some differences between the slow and fast algorithms are consistently the same number, as if there was a constant added. This constant value is, I think, the 'garbage' value that the sum picks up when it runs past the valid indices.


Former-commit-id: e2bac67c522f985804c8f53c6de6eaba62543c5a