Update issue 1356
Fixed sloppy logic in old fetch algorithm's loop where the error in depths was comming from. The loop aggregated one value too many, which was the reason for the discrepancy of ~1.7e-5. Now the depths and distances agree for both versions of the algorithm down to the 15th significant digit.


Former-commit-id: a9c51fb46c8a73bc302e789ced44d3d30b043f94