Update issue 1376
When I actually got down to it, I realized that almost all of the
biophysical function is accomplished by two separate functions, and that
most of the extra stuff is just looping through the various species,
and then operating on rasters summed across species.  I concluded that
it would be fine to have the whole logic of the model in a single
function, which I have now done.  Still need to ensure that tests pass.


Former-commit-id: 6972b873598c2b91c240c94bcd2ff98e86ec29c1