Update issue 1527
Updating the call to mkdtemp() so that the parameter provided by the
user is passed to tempfile.mkdtemp() so that the new temp folder that is
created is made inside of the user-defined folder.

Under the previous approach (where the user-defined folder would be used
as the folder for the gaussian filter's mem-mapped files), the user was
being asked to ensure that the folder exists on disk before calling
gaussian_filer_dataset.  This assumption is unreasonable in this
circumstance because gaussian_filter_dataset() also _deletes_ whatever
temp folder is used in the function.  If the folder used is a brand-new
temp folder, no problem.  But if the user wants to define the temp
folder used, the dev would need to recreate the folder after _each_ call
to gaussian_filter_dataset.

I think the modification this changeset introduces is a good compromise,
as it allows the user to define where the folder will exist (such as on
an external disk drive, if desired), while still ensuring that all temp
files are created in a new folder _inside_ of the user-defined folder.
Then, at the end of the function, this new folder is deleted without
issue.

Rich, I'm sorry I didn't have an opportunity to ask you about this in
person ... you were plugged in most of the afternoon.  Feel free to
reopen!


Former-commit-id: 4096287b23feb4c5f2d3a6ac4aa57b8356e23797