Update issue 1189
As Greg Guannel pointed out, I didn't use equation 10 in the UG to compute wave power. There was some code added previously that didn't work and was left commented out. It was in the wind exposure function. Moved that code in compute_wave_exposure, and started to uncomment the lines that compute the variables H and T needed to compute wave power. So far, the computation of H is running smoothly. Next is the computation of T.


Former-commit-id: 1dc117033457d4c58a485ced33d27e86b2f5fbac