Update issue 1487
Moved the loop for the base layers back in the execute function to include more computational steps. I had to [temporarily] give up the post-processing phases needed for the layers used in the computation of the coastal vulnerability index. Will add them later, possibly through another field in the layer_information (soon to be renamed something like output information).


Former-commit-id: 39480db55558b1c2bb0531b1890cf22619882564