Update issue 1487
Modified the preconditions from a list to a list of lists so I can test necessary and optional keywords in the same construct. Basically, all keywords in an inner list have to exist in the arguments dictionary for the list to evaluate to True, and the outer list evaluate to true if at least one inner list is true.


Former-commit-id: f71722f78bc446796b94a995690ef7ef498e1b62