Update issue 1487
Committing the code that used to be in the if statements and that is now in the main loop in execute-core. Although the code brakes during execution because the refactoring is not complete, I want to commit this because it's already a lot of code changes. I also added some comments, following Rich's recommendations.


Former-commit-id: 5bf0b6732195888ab2db75ea97eea6b5e79451e9