Update issue 1251
Adding proper masking to valuation outputs.

Since valuation and biophysical are combined I can now open a raster output from
the biophysical run and use that to mask against the valuation outputs. This is
helpful because the biophysical outputs are properly masked to depth, distance
and other parameters. Removing masking by land poly from valuation because that
is done in biophysical, whose outputs we are now using to mask valuation
outputs.


Former-commit-id: e9ca821063ac808b1a09859bacbb0e07ededf297