Update issue 1494
Adding the global wind energy parameter json file located under wind energy.
Edited setup.py to search for this file and place it in
invest_natcap/wind_energy/ .

It will no longer be located in the input folder.

Still need to add the file as a obvious global element in wind_energy_core


Former-commit-id: 45cff21ac0be3378455927ee8b20071220ed7b45