Update issue 1548
So I think I have actually fixed the issue here.  We cannot reasonably
assume that all CSVs will have double-quoted strings and that it will be
comma-delimited.  Given the CSV module's tendency to return strings
instead of floats, we cannot assume that CSV will return the correct
values either.

So here's the solution, as of this commit:  Loop through all values in a
CSV table that is being parsed by the CSVDriver class for an instance of
TableHandler.  For each valyue, try to cast it to a float.  If it fails,
return the value.  Otherwise, return the float.

All validator, pollination and nutrient retention (smoke) tests pass with
this patch.


Former-commit-id: 0acb9cc91b19394a18d891a844a57e321fb01d8b