Update issue 1540
Rewrote how the try/except block is setup around the AOI argument. It is now
setup as a try/except/else block.

Also moved the creation of the wind energy points from data file into both the
try and except situations so that I could decide what the uri path should be
named.


Former-commit-id: 1ff2d9600c5dce19e72f0748ef35ea4edccfff69