Update issue 1544
I'm adding some new functionality to the UI:  all elements now have two
new functions: get_element_state() and set_element_state().  The default
behavior is to use the existing value() and setValue() functions, just
as we had before.  But now if we have newer, more complex elements, we
can save more information to a dictionary.

This functionality is strictly used for the loading and saving of
lastrun information.


Former-commit-id: e02fc23276c3d2e8f5991dbe272ee3fb0911bb2e